pyproject.toml
[tool.poetry]
name = "django-browserless"
version = "0.1.0"
description = "Genreate PDFs from URLs using browserless"
authors = ["Sylvain Gable <sylvain@vingtcinq.io>"]
license = "MIT License"
readme = "README.md"
packages = [
    { include = "django_browserless" },
]

[tool.poetry.dependencies]
attrs = "^23.2.0"
python = "^3.8"
Django = "^4.2"
djangorestframework = "^3.12.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pytest-clarity = "^1.0.1"

[tool.ruff]
target-version = "py312"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
